backend/requirements/requirements.txt
# =============================================================================
# Price Watch - Production Requirements
# =============================================================================

# Django & REST Framework
Django>=5.0
djangorestframework
django-filter
django-cors-headers
psycopg[binary]
python-dotenv
celery
redis
requests
pydantic
structlog
